requirements.txt
openai
streamlit
pypdf
pypdfium2
plotly
pandas
Pillow
python-docx
reportlab
pytest
